requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gpiod==1.5.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
RPi.GPIO2==0.3.0a3
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
